# docs/openapi.yaml
...
openapi: 3.1.1
info:
  title: AR.IO Gateway
  version: 0.0.1
  x-logo:
    url: 'https://ar.io/wp-content/uploads/2022/02/ar-io-FINAL-LOGO-SHADOW.png'
  description: |
    AR.IO Gateways are the interface between users and the Permaweb. Each Gateway acts like a “Permaweb Service
    Provider” and supports multiple, value-added, low-trust services for users and applications.

    The AR.IO HTTP API is served by all AR.IO Gateways, and is used by all Arweave clients looking to read, write and query data.  It supports the following capabilities
    - seeding layer one transactions to the greater Arweave network
    - caching and serving chain and transaction data
    - resolving friendly names to transaction data with the Arweave Name System
    - indexing and querying layer one and two transactions
    - proxy for Arweave nodes serving the Arweave HTTP API

    ## Errors

    The API uses standard HTTP status codes to indicate the success or failure of the API call.

    ## Authentication

    All inbound transactions to the Arweave network must contain a valid signature from an Arweave wallet with appropriate amount of AR tokens.  At this time, clients are not required to sign their requests for outbound data from the Arweave network.
  license:
    name: GNU Affero General Public License v3.0
    identifier: AGPL-3.0-only
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: apiToken
      description: ADMIN_API_KEY set in your .env file.
...
  responses:
...
    ServerError:
      description: Internal Server Error
      content:
        text/plain:
          schema:
            type: string
            example: 'Internal Server Error'
paths:
...
  '/ar-io/rate-limit/ip/{ip}':
    get:
      tags: [Rate Limiting]
      summary: Query IP bucket balance
      description: |
        Get the current rate limit token balance for an IP address. Returns both regular
        (refilling) and paid tokens. Buckets are created on first request and refill over time.

        **Token System:**
        - 1 token = 1 KiB (1,024 bytes)
        - Regular tokens: Automatically refill based on configured rate (e.g., 20 KB/s)
        - Paid tokens: Acquired via x402 payments, never refill, consumed after regular tokens
      operationId: getRateLimitIpBalance
      parameters:
        - name: ip
          in: path
          required: true
          schema:
            type: string
            format: ipv4 or ipv6
          description: IP address (IPv4, IPv6, or IPv4-mapped IPv6 format)
          example: '192.168.1.100'
      responses:
        '200':
          description: Successfully retrieved IP bucket balance
          content:
            application/json:
              schema:
                type: object
                properties:
                  ip:
                    type: string
                    description: IP address
                    example: '192.168.1.100'
                  tokens:
                    type: number
                    description: Regular tokens (refilling)
                    example: 95000
                  paidTokens:
                    type: number
                    description: Paid tokens (non-refilling)
                    example: 50000
                  capacity:
                    type: number
                    description: Maximum regular token capacity
                    example: 100000
                  refillRate:
                    type: number
                    description: Refill rate in tokens per second
                    example: 20
                  lastRefill:
                    type: number
                    format: int64
                    description: Last refill timestamp (Unix milliseconds)
                    example: 1735689600000
                required:
                  - ip
                  - tokens
                  - paidTokens
                  - capacity
                  - refillRate
                  - lastRefill
        '400':
          description: Invalid IP address format
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Invalid IP address format'
        '404':
          description: Bucket not found (created on first request)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Bucket not found'
                  message:
                    type: string
                    example: 'No rate limit bucket exists for this IP. Buckets are created on first request.'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [Rate Limiting]
      summary: Top up IP bucket
      description: |
        Add paid tokens to an IP bucket via x402 payment or admin authentication.

        **Authentication Modes:**

        1. **x402 Payment (Public):**
           - Provide X-Payment header with payment data
           - Tokens calculated from payment amount
           - 10x capacity multiplier applied
           - Example: $0.10 payment → ~976 KB equivalent → ~976,000 tokens

        2. **Admin (Private/Testing):**
           - Provide Authorization: Bearer header with admin API key
           - Specify exact token count and type in request body
           - No multiplier applied (raw token count)
           - Only supports paid tokens currently
      operationId: topUpRateLimitIpBucket
      parameters:
        - name: ip
          in: path
          required: true
          schema:
            type: string
            format: ipv4 or ipv6
          description: IP address
          example: '192.168.1.100'
        - name: X-Payment
          in: header
          required: false
          schema:
            type: string
            format: base64
          description: |
            Base64-encoded x402 payment data (for x402 payment mode).
            Mutually exclusive with Authorization header.
      security:
        - {}
        - bearerAuth: []
      requestBody:
        description: |
          Required for admin authentication mode. Omit for x402 payment mode.
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                tokens:
                  type: number
                  minimum: 1
                  description: Number of tokens to add (admin mode only)
                  example: 50000
                tokenType:
                  type: string
                  enum: [paid]
                  description: Type of tokens (admin mode only, currently only "paid" supported)
                  example: 'paid'
              required:
                - tokens
                - tokenType
      responses:
        '200':
          description: Successfully topped up IP bucket
          content:
            application/json:
              schema:
                type: object
                properties:
                  ip:
                    type: string
                    example: '192.168.1.100'
                  tokens:
                    type: number
                    example: 95000
                  paidTokens:
                    type: number
                    example: 150000
                  capacity:
                    type: number
                    example: 100000
                  refillRate:
                    type: number
                    example: 20
                  lastRefill:
                    type: number
                    format: int64
                    example: 1735689600000
                  topUp:
                    type: object
                    description: Information about the top-up operation
                    properties:
                      tokensAdded:
                        type: number
                        description: Number of tokens added
                        example: 100000
                      paymentAmount:
                        type: string
                        description: Payment amount in USDC atomic units (x402 mode only)
                        example: '1000000'
                      multiplierApplied:
                        type: number
                        description: Capacity multiplier applied (10 for x402, 1 for admin)
                        example: 10
                      tokenType:
                        type: string
                        description: Type of tokens added (admin mode only)
                        example: 'paid'
                    required:
                      - tokensAdded
                      - multiplierApplied
        '400':
          description: Invalid IP format or invalid request body
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Invalid tokens value'
                  message:
                    type: string
                    example: 'tokens must be a positive number'
        '401':
          description: Missing or invalid authentication
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Unauthorized'
                  message:
                    type: string
                    example: 'Provide X-Payment header for x402 payment or Authorization: Bearer header for admin access'
        '402':
          description: Payment verification or settlement failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Payment failed'
                  message:
                    type: string
                    example: 'Payment verification failed: insufficient amount'
        '500':
          $ref: '#/components/responses/ServerError'
...